+++ e2e/pages/administration/jobs/scheduled/scheduled.readonly.spec.js
@@ -16,10 +16,12 @@ describe('when an ONC-Staff user is logged in', () => {
   beforeEach(function () {
     loginComponent.logInWithEmail('oncstaff');
     hooks.open('#/administration/jobs/scheduled');
+    hooks.waitForSpinnerToDisappear();
   });
 
   it('should see the right set of scheduled jobs', () => {
     const expected = new Set(['All Broken Surveillance Rules Report','Developer Access Report','Inherited Certification Status Errors Report','Listing Validation Email Report','Overnight Broken Surveillance Rules Report','Pending "Change Request" Report','Questionable Activity Report','Questionable URL Report','Real World Testing Email Report','Summary Statistics Email','Trigger Developer Ban Notification']);
+    browser.waitUntil(() => scheduled.scheduledJobTable.isDisplayed());
     assert.equal(scheduled.scheduledJobRows.length,expected.size);
     // get the existing jobs into a de-duplicated array of jobs
     let rows = [...new Set(scheduled.scheduledJobRows.map(row => row.getText()))];