+++ src/app/pages/surveillance/reporting/quarter.component.js
@@ -43,6 +43,10 @@ export const SurveillanceReportQuarterComponent = {
       this.onSave({report: this.report});
     }
 
+    can (action) {
+      return action === 'delete' && this.hasAnyRole(['ROLE_ADMIN', 'ROLE_ACB']);
+    }
+
     cancel () {
       this.report = angular.copy(this.backup.report);
       this.onCancel();
@@ -52,16 +56,38 @@ export const SurveillanceReportQuarterComponent = {
       this.takeAction({report: this.report, action: 'delete'});
     }
 
+    takeActionBarAction (action) {
+      switch (action) {
+      case 'cancel':
+        this.cancel();
+        break;
+      case 'delete':
+        this.delete();
+        break;
+      case 'save':
+        this.save();
+        break;
+      //no default
+      }
+    }
+
     generateReport () {
       let that = this;
       this.networkService.generateQuarterlySurveillanceReport(this.report.id)
-        .then(results => {
-          let name = results.user.friendlyName ? results.user.friendlyName : results.user.fullName;
-          let email = results.user.email;
+        .then(response => {
+          let name = response.job.jobDataMap.user.friendlyName || response.job.jobDataMap.user.fullName;
+          let email = response.job.jobDataMap.user.email;
           that.toaster.pop({
             type: 'success',
             title: 'Report is being generated',
             body: `Quarterly Surveillance report is being generated, and will be emailed to ${name} at ${email} when ready.`,
+          });
+        }, error => {
+          let message = error.data.error;
+          that.toaster.pop({
+            type: 'error',
+            title: 'Report could not be generated',
+            body: message,
           });
         });
     }